+++ crates/forge-runner/src/test_case_summary.rs
@@ -1,11 +1,16 @@
 use crate::compiled_runnable::TestCaseRunnable;
 use crate::expected_result::{ExpectedPanicValue, ExpectedTestResult};
+use crate::gas::check_available_gas;
+use crate::trace_data::CallTrace;
 use cairo_felt::Felt252;
 use cairo_lang_runner::short_string::as_cairo_short_string;
 use cairo_lang_runner::{RunResult, RunResultValue};
 use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::UsedResources;
+use cheatnet::state::CallTrace as InternalCallTrace;
 use num_traits::Pow;
+use std::cell::RefCell;
 use std::option::Option;
+use std::rc::Rc;
 
 #[derive(Debug, PartialEq, Clone, Default)]
 pub struct GasStatistics {
@@ -52,6 +57,7 @@ pub struct FuzzingStatistics {
 pub trait TestType {
     type GasInfo: std::fmt::Debug + Clone;
     type TestStatistics: std::fmt::Debug + Clone;
+    type TraceData: std::fmt::Debug + Clone;
 }
 
 #[derive(Debug, PartialEq, Clone)]
@@ -59,6 +65,7 @@ pub struct Fuzzing;
 impl TestType for Fuzzing {
     type GasInfo = GasStatistics;
     type TestStatistics = FuzzingStatistics;
+    type TraceData = ();
 }
 
 #[derive(Debug, PartialEq, Clone)]
@@ -66,6 +73,7 @@ pub struct Single;
 impl TestType for Single {
     type GasInfo = u128;
     type TestStatistics = ();
+    type TraceData = CallTrace;
 }
 
 /// Summary of running a single test case
@@ -85,6 +93,8 @@ pub enum TestCaseSummary<T: TestType> {
         used_resources: UsedResources,
         /// Statistics of the test run
         test_statistics: <T as TestType>::TestStatistics,
+        /// Test trace data
+        trace_data: <T as TestType>::TraceData,
     },
     /// Test case failed
     Failed {
@@ -150,6 +160,7 @@ impl TestCaseSummary<Fuzzing> {
                 gas_info: _,
                 used_resources: _,
                 test_statistics: (),
+                trace_data: _,
             } => {
                 let runs = results.len();
                 let gas_usages: Vec<u128> = results
@@ -167,6 +178,7 @@ impl TestCaseSummary<Fuzzing> {
                     gas_info: GasStatistics::new(&gas_usages),
                     used_resources: UsedResources::default(),
                     test_statistics: FuzzingStatistics { runs },
+                    trace_data: (),
                 }
             }
             TestCaseSummary::Failed {
@@ -196,19 +208,24 @@ impl TestCaseSummary<Single> {
         arguments: Vec<Felt252>,
         gas: u128,
         used_resources: UsedResources,
+        call_trace: &Rc<RefCell<InternalCallTrace>>,
     ) -> Self {
         let name = test_case.name.to_string();
         let msg = extract_result_data(&run_result, &test_case.expected_result);
         match run_result.value {
             RunResultValue::Success(_) => match &test_case.expected_result {
-                ExpectedTestResult::Success => TestCaseSummary::Passed {
-                    name,
-                    msg,
-                    arguments,
-                    test_statistics: (),
-                    gas_info: gas,
-                    used_resources,
-                },
+                ExpectedTestResult::Success => {
+                    let summary = TestCaseSummary::Passed {
+                        name,
+                        msg,
+                        arguments,
+                        test_statistics: (),
+                        gas_info: gas,
+                        used_resources,
+                        trace_data: CallTrace::from(call_trace.borrow().clone()),
+                    };
+                    check_available_gas(&test_case.available_gas, summary)
+                }
                 ExpectedTestResult::Panics(_) => TestCaseSummary::Failed {
                     name,
                     msg,
@@ -239,6 +256,7 @@ impl TestCaseSummary<Single> {
                         test_statistics: (),
                         gas_info: gas,
                         used_resources,
+                        trace_data: CallTrace::from(call_trace.borrow().clone()),
                     },
                 },
             },