+++ crates/forge/src/main.rs
@@ -102,6 +102,10 @@ struct TestArgs {
     /// Run tests that failed during the last run
     #[arg(long)]
     rerun_failed: bool,
+
+    /// Save execution traces of all test which have passed and are not fuzz tests
+    #[arg(long)]
+    save_trace_data: bool,
 }
 
 fn validate_fuzzer_runs_value(val: &str) -> Result<u32> {
@@ -134,6 +138,7 @@ fn combine_configs(
     fuzzer_runs: Option<u32>,
     fuzzer_seed: Option<u64>,
     detailed_resources: bool,
+    save_trace_data: bool,
     forge_config: &ForgeConfig,
 ) -> RunnerConfig {
     RunnerConfig::new(
@@ -146,6 +151,7 @@ fn combine_configs(
             .or(forge_config.fuzzer_seed)
             .unwrap_or_else(|| thread_rng().next_u64()),
         detailed_resources,
+        save_trace_data || forge_config.save_trace_data,
     )
 }
 
@@ -229,6 +235,7 @@ fn test_workspace(args: TestArgs) -> Result<bool> {
                     args.fuzzer_runs,
                     args.fuzzer_seed,
                     args.detailed_resources,
+                    args.save_trace_data,
                     &forge_config,
                 ));
                 let runner_params = Arc::new(RunnerParams::new(contracts, env::vars().collect()));
@@ -312,6 +319,7 @@ mod tests {
             None,
             None,
             false,
+            false,
             &Default::default(),
         );
         let config2 = combine_configs(
@@ -319,6 +327,7 @@ mod tests {
             false,
             None,
             None,
+            false,
             false,
             &Default::default(),
         );
@@ -337,6 +346,7 @@ mod tests {
             None,
             None,
             false,
+            false,
             &Default::default(),
         );
         assert_eq!(
@@ -346,6 +356,7 @@ mod tests {
                 false,
                 FUZZER_RUNS_DEFAULT,
                 config.fuzzer_seed,
+                false,
                 false,
             )
         );
@@ -358,13 +369,22 @@ mod tests {
             fork: vec![],
             fuzzer_runs: Some(1234),
             fuzzer_seed: Some(500),
+            save_trace_data: true,
         };
         let workspace_root: Utf8PathBuf = Default::default();
 
-        let config = combine_configs(&workspace_root, false, None, None, true, &config_from_scarb);
+        let config = combine_configs(
+            &workspace_root,
+            false,
+            None,
+            None,
+            true,
+            false,
+            &config_from_scarb,
+        );
         assert_eq!(
             config,
-            RunnerConfig::new(workspace_root, true, 1234, 500, true)
+            RunnerConfig::new(workspace_root, true, 1234, 500, true, false)
         );
     }
 
@@ -377,6 +397,7 @@ mod tests {
             fork: vec![],
             fuzzer_runs: Some(1234),
             fuzzer_seed: Some(1000),
+            save_trace_data: false,
         };
         let config = combine_configs(
             &workspace_root,
@@ -384,12 +405,13 @@ mod tests {
             Some(100),
             Some(32),
             true,
+            true,
             &config_from_scarb,
         );
 
         assert_eq!(
             config,
-            RunnerConfig::new(workspace_root, true, 100, 32, true)
+            RunnerConfig::new(workspace_root, true, 100, 32, true, true)
         );
     }
 }