+++ CHANGELOG.md
@@ -9,14 +9,44 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Forge
 
+#### Changed
+
+- Gas estimation is now aligned with the Starknet v0.13
+
+## [0.16.0] - 2024-01-26
+
+### Forge
+
+#### Added
+- Bump to cairo 2.5.0
+
+#### Changed
+
+- `SafeDispatcher`s usages need to be tagged with `#[feature("safe_dispatcher)]` (directly before usage), see [the shamans post](https://community.starknet.io/t/cairo-v2-5-0-is-out/112807#safe-dispatchers-15)
+
+## [0.15.0] - 2024-01-24
+
+### Forge
+
 #### Added
 
 - `--detailed-resources` flag for displaying additional info about used resources
 - `store` and `load` cheatcodes
+- `--save-trace-data` flag to `snforge test` command. Traces can be used for profiling purposes.
+
+#### Changed
+
+- `available_gas` attribute is now supported (Scarb >= 2.4.4 is required)
 
 #### Fixed
 
 - Error message for tests that should panic but pass
+
+### Cast
+
+#### Changed
+
+- the 'pending' block is used instead of 'latest' as the default when obtaining the nonce
 
 ## [0.14.0] - 2024-01-11